# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/businesses": {
      "post": {
        "operationId": "BusinessController_create",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new business",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/{id}": {
      "put": {
        "operationId": "BusinessController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a business",
        "tags": [
          "Business"
        ]
      },
      "delete": {
        "operationId": "BusinessController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a business",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/search": {
      "get": {
        "operationId": "BusinessController_search",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "example": "restaurant",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Search businesses by keyword",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/filter": {
      "post": {
        "operationId": "BusinessController_filter",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "orderDirection",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Filter businesses with advanced criteria",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/account/{accountId}": {
      "get": {
        "operationId": "BusinessController_getByAccount",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get businesses by account ID",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/proximity/{lat}/{lng}/{categoryId}": {
      "get": {
        "operationId": "BusinessController_getByProximity",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "path",
            "schema": {
              "example": -23.5505,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "path",
            "schema": {
              "example": -46.6333,
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "maxDistance",
            "required": false,
            "in": "query",
            "schema": {
              "example": 5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get businesses by location proximity",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/details/{id}": {
      "get": {
        "operationId": "BusinessController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get business by ID",
        "tags": [
          "Business"
        ]
      }
    },
    "/businesses/details/{id}/increment-views": {
      "post": {
        "operationId": "BusinessController_incrementViews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Increment business view count",
        "tags": [
          "Business"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "CategoriesController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all categories with subcategories",
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/categories/{id}": {
      "put": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a category",
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/comments/business/{businessId}": {
      "get": {
        "operationId": "CommentsController_getByBusiness",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get comments by business ID",
        "tags": [
          "Comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Circuito BR 116 API",
    "description": "API documentation for Circuito BR 116",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local environment"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production environment"
    }
  ],
  "components": {
    "schemas": {
      "BusinessResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Business UUID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "accountId": {
            "type": "number",
            "description": "Account ID",
            "example": 1
          },
          "locationId": {
            "type": "number",
            "description": "Location ID",
            "example": 1
          },
          "locationPretty": {
            "type": "string",
            "description": "Location name",
            "example": "São Paulo, SP"
          },
          "locationLat": {
            "type": "number",
            "description": "Latitude coordinate",
            "example": -23.5505
          },
          "locationLong": {
            "type": "number",
            "description": "Longitude coordinate",
            "example": -46.6333
          },
          "urlImage": {
            "type": "string",
            "description": "Business Logo URL",
            "example": "https://example.com/image.jpg"
          },
          "title": {
            "type": "string",
            "description": "Business title",
            "example": "Amazing Restaurant"
          },
          "description": {
            "type": "string",
            "description": "Business description",
            "example": "The best food in town"
          },
          "views": {
            "type": "number",
            "description": "View count",
            "example": 150
          },
          "isVerified": {
            "type": "boolean",
            "description": "Verification status",
            "example": false
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "+5511999999999"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "contact@business.com"
          },
          "address": {
            "type": "string",
            "description": "Physical address",
            "example": "Rua Example, 123"
          },
          "promotedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Promotion date"
          },
          "whatsapp": {
            "type": "string",
            "description": "WhatsApp number",
            "example": "+5511999999999"
          },
          "facebook": {
            "type": "string",
            "description": "Facebook profile URL"
          },
          "instagram": {
            "type": "string",
            "description": "Instagram handle"
          },
          "tiktok": {
            "type": "string",
            "description": "TikTok handle"
          },
          "categoryId": {
            "type": "number",
            "description": "Category ID",
            "example": 1
          },
          "classification": {
            "type": "string",
            "description": "Classification code",
            "example": "A1"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "accountId",
          "locationId",
          "locationPretty",
          "locationLat",
          "locationLong",
          "urlImage",
          "title",
          "description",
          "views",
          "isVerified",
          "categoryId",
          "classification",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Business title",
            "example": "Amazing Restaurant"
          },
          "description": {
            "type": "string",
            "description": "Business description",
            "example": "The best food in town"
          },
          "locationPretty": {
            "type": "string",
            "description": "Location name",
            "example": "São Paulo, SP"
          },
          "locationId": {
            "type": "number",
            "description": "Location ID",
            "example": 1
          },
          "locationLat": {
            "type": "number",
            "description": "Latitude coordinate",
            "example": -23.5505
          },
          "locationLong": {
            "type": "number",
            "description": "Longitude coordinate",
            "example": -46.6333
          },
          "categoryId": {
            "type": "number",
            "description": "Category ID",
            "example": 1
          },
          "classification": {
            "type": "string",
            "description": "Classification code",
            "example": "A1"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "+5511999999999"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "contact@business.com"
          },
          "address": {
            "type": "string",
            "description": "Physical address",
            "example": "Rua Example, 123"
          },
          "whatsapp": {
            "type": "string",
            "description": "WhatsApp number",
            "example": "+5511999999999"
          },
          "facebook": {
            "type": "string",
            "description": "Facebook profile URL"
          },
          "instagram": {
            "type": "string",
            "description": "Instagram handle"
          },
          "tiktok": {
            "type": "string",
            "description": "TikTok handle"
          }
        },
        "required": [
          "title",
          "locationPretty",
          "locationLat",
          "locationLong",
          "categoryId"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Business title",
            "example": "Amazing Restaurant"
          },
          "description": {
            "type": "string",
            "description": "Business description",
            "example": "The best food in town"
          },
          "locationPretty": {
            "type": "string",
            "description": "Location name",
            "example": "São Paulo, SP"
          },
          "locationLat": {
            "type": "number",
            "description": "Latitude coordinate",
            "example": -23.5505
          },
          "locationLong": {
            "type": "number",
            "description": "Longitude coordinate",
            "example": -46.6333
          },
          "categoryId": {
            "type": "number",
            "description": "Category ID",
            "example": 1
          },
          "classification": {
            "type": "string",
            "description": "Classification code",
            "example": "A1"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "+5511999999999"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "contact@business.com"
          },
          "address": {
            "type": "string",
            "description": "Physical address",
            "example": "Rua Example, 123"
          },
          "whatsapp": {
            "type": "string",
            "description": "WhatsApp number",
            "example": "+5511999999999"
          },
          "facebook": {
            "type": "string",
            "description": "Facebook profile URL"
          },
          "instagram": {
            "type": "string",
            "description": "Instagram handle"
          },
          "tiktok": {
            "type": "string",
            "description": "TikTok handle"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Verification status"
          },
          "views": {
            "type": "number",
            "description": "View count"
          }
        }
      },
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "List of entities for the current page.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "currentPage": {
            "type": "number",
            "example": 1,
            "description": "Current page number."
          },
          "totalElements": {
            "type": "number",
            "example": 100,
            "description": "Total number of elements."
          },
          "nextPage": {
            "type": "number",
            "example": 2,
            "description": "Next page number, or null if last page."
          },
          "lastPage": {
            "type": "number",
            "example": 10,
            "description": "Last page number."
          },
          "firstPage": {
            "type": "number",
            "example": 1,
            "description": "First page number."
          },
          "previousPage": {
            "type": "number",
            "example": 1,
            "description": "Previous page number, or null if first page."
          }
        },
        "required": [
          "data",
          "currentPage",
          "totalElements",
          "nextPage",
          "lastPage",
          "firstPage",
          "previousPage"
        ]
      },
      "FilterBusinessDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "Search query"
          },
          "categories": {
            "description": "Category IDs to filter by",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "locationIds": {
            "description": "Location IDs to filter by",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "classifications": {
            "description": "Classification codes to filter by",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "promotedOnly": {
            "type": "boolean",
            "description": "Filter only promoted businesses"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Filter only verified businesses"
          },
          "minViews": {
            "type": "number",
            "description": "Minimum view count"
          },
          "maxViews": {
            "type": "number",
            "description": "Maximum view count"
          },
          "accountId": {
            "type": "number",
            "description": "Account ID to filter by"
          }
        }
      },
...
    }
  }
}